+++ includes/Render/FieldBlocks.php
@@ -224,6 +224,7 @@ final class FieldBlocks {
 				'image'   => (bool) $field['image'],
 				'bullets' => (bool) $field['bullets'],
 				'filters' => (bool) $field['filters'],
+				'link'    => (bool) $field['link'],
 				'columns' => $columns,
 			);
 		}